Api allows for query parameters to filter
The only typesafe way I could find to implement filters was to use a CheckFilter and then check for that in the CheckContext. The CheckContext is required because of multi-threading issues and an object needed to be passed into CheckRunner so the context could be set anyway. Hopefully this api change will make it easier for checks to call other checks with objects that are already setup like a Socket or something else that can extend a check.